pyproject.toml
[tool.poetry]
name = "weathermate"
version = "0.1.0"
description = ""
authors = ["bluewarn <n5lp97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ttkbootstrap = "^1.10.1"
tkdesigner = "^1.0.7"
matplotlib = "^3.8.4"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
pandas = "^2.2.2"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
